requirements.txt
flask==1.1.2
werkzeug==1.0.1
tqdm==4.55.0
Pillow==8.0.1
opencv-python==4.4.0.46
numpy==1.19.4
tensorflow==2.2.0
